backend/requirements.txt
Flask==2.2.5
gunicorn==21.2.0
redis==4.6.0
python-dotenv==1.0.0
azure-storage-blob==12.17.0
flask-cors==4.0.0
Werkzeug==2.2.3
